Move util.unpack_from_dynamodb to models.unpack_from_dynamodb
